Perfect timer
- *Time -> *Duration for more consistent variable naming
- ozCore
  * Log::operator<<([u]long)
- ozEngine
  * Application uses perfect timing (see Timer)
- common
  * Timer has variable (real) tick duration, so sum of all tick durations in a
    seconds always yields exactly one second; for game time tick duration is
    still a compile-time constant

+++ src/ozCore/Instant.cc
@@ -34,7 +34,7 @@ Instant Instant::now()
   struct timespec ts;
   clock_gettime(CLOCK_MONOTONIC, &ts);
 
-  return Instant(ts.tv_sec * 1000000000 + ts.tv_nsec);
+  return Instant(int64(ts.tv_sec) * 1000000000 + int64(ts.tv_nsec));
 }
 
 }
